[ADD] 라이브서비스와 주고 받는 메시지를 unsigned 16비트로 변경

+++ lib/models/protocols/const.dart
@@ -27,7 +27,7 @@ enum DustyAction {
   final int code;
   const DustyAction(this.code);
 
-  ByteBuffer encode() => Uint8List.fromList([code]).buffer;
+  ByteBuffer encode() => Uint16List.fromList([code]).buffer;
 }
 
 enum DustyState {
